+++ degu/server.py
@@ -30,7 +30,6 @@ from os import path
 
 from .base import (
     TYPE_ERROR,
-    EmptyLineError,
     makefiles,
     read_preamble,
     parse_headers,
@@ -421,7 +420,7 @@ class Handler:
         request = self.environ.copy()
         try:
             request.update(self.build_request())
-        except (EmptyLineError, socket.timeout):
+        except (ConnectionError, socket.timeout):
             return self.shutdown()
         if request['method'] not in {'GET', 'PUT', 'POST', 'DELETE', 'HEAD'}:
             return self.write_status_only(405, 'Method Not Allowed')